Ensure replication_network is not required for swift
* Adjust config location variables for account, container, object
variables, to name them more clearly.
* Adjust config placement order, to happen after init scripts are
created so that the service restart handler will work appropriately.
* Ensure the init scripts for the replicator service are created
appropriately whether a dedicated replication_network is or isn't used.
* Ensure the dedicated replication network configuration is only created
when replication_network is specified and different to storage_network.
* Ensure the appropriate configuration section in the server.conf is
only added when the replication_network is specified and different to
the storage_network.

Change-Id: I4b204a974bb0a217f5222b6aa0fa36aa8c23b999
Closes-Bug: #1427599