Distribute rings to build dir on all swift_hosts
This prevents a situation where swift_hosts[0] changes as a host, or is
removed (meaning a new swift_hosts[0]) which would prevent the rings
from being built. Additionally it will act as a backup for the ring
build files.

To clarify the purpose of the rings directory, the directory was renamed
to "ring_build_files", clarifying that the active rings are in
/etc/swift/

Change-Id: Id878d219504db195353286b28d7c038e5e659263
Closes-Bug: #1444529