Fix rsync service restart in os_swift
The rsync service is currently restarted using two handlers, one to stop
the service and a second to start it. There is not a sufficient delay
between the two task and so the rsync pid has not been removed before
the attempt is made to start the service.

This commit replaces the two handlers with a single one that will do the
restart in one go.

Change-Id: I8ed4630da1add7205552b6ec731a143dbe45112b
Closes-bug: 1538649