Merge os-swift-sync role into os-swift
The repo split out has meant that having multiple roles for swift has
become inconvenient. We can merge the os-swift-sync role back into
os-swift and utilise variables to ensure that the swift-install,
swift-setup and swift-sync playbooks work in the same way.

This will avoid code duplication and the requirement for duplicated
changes to happen to both repositories.

Change-Id: If01d8a3aa29fc7637aecb92e61929523af08103d