Remove broken package "pywbem"
This pacakge is causing system breakage due to a new version being
released that is otherwise incompatible with requirements as found
within OpenStack. After reviewing the package usage and the reason
it's been included but the only thing I could find that actually
uses this resource was cinder and only when using the EMC VMAX
driver: http://paste.openstack.org/show/491185/ -- being that this
package is not even listed in global requirements and is not in
support of anything we could say is a common backend within Cinder
that we know we're currently supporting I am of the opinion that
this package should be removed.

Change-Id: I152fa0130e64ad5756b413ed8f9714d28e75d0bb
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>