Limit local tasks to local hosts
We want to improve the gate speed by running the swift-role once on AIO,
since we don't need to split the sync tasks out when not using MR swift.
This means for MR swift we need to differentiate between the tasks that
should be run on remote hosts and those that should be run on local
hosts.

Change-Id: I0a8436ebbab0226f6c4fb499c938f760429aad39