Clean up deprecation warning in swift proxy
The following warning appears in the swift proxy logs:
"Using the in-process token cache is deprecated as of the 4.2.0
release and may be removed in the 5.0.0 release or the 'O'
development cycle. The in-process cache causes inconsistent results
and high memory usage."

Specifying the location of a memcached server will not only clean up
a deprecation warning in the logs but should provide a performance
improvement.

This is caused by the absence of a memcached_servers key in the
proxy configuration. While this variable is defined as a group var
in the integrated release we insulate against the absense of this
value in stand-alone role usage or Swift Stand-Alone deploys which
may not include memcached yet to prevent upgrade impacts.

Change-Id: I0a16ef0030d1425f52b6c07214d7c8effef51c13