Skip unavailable hosts when distributing keys
The 'Create authorized keys file from host vars' task assumes that every
host within the swift_all and swift_remote_all groups is available. If
any host within those groups is offline, excluded through a limit, or
has otherwise failed an earlier task, it will lead to this task failing
for every host.

Add a when clause checking for the existence of the swift_pubkey
hostvar so that the task and remainder of the role can continue when a
swift host is unavailable.

Change-Id: Iee8d5557ae5965d75fd433009796c26324f5dc80