Force Ansible to use dynamic includes
Ansible 2.1.1 introduces a regression in the way conditional
includes are handled which results in every task in the
included file being evaluated even if the condition for the
include is not met. This extends the run time significantly
for a deployment.

This patch forces all conditional includes which are not
inventory based to be dynamic.

Change-Id: Ied462a7b79cb2d6615e110484ea734e136f59034
Related-Bug: https://github.com/ansible/ansible/issues/17687