Simplify networking for MR swift
We're already testing dedicated replication networks in the base swift
build, lets reduce build time and simplify the networking for the Multi
Region swift build.

Additionally, the infra host needs to be able to route back to the other
hosts in the Multi-Region solution, as it is now the single Infra Host.

Change-Id: If6fd211cd95c4f068293cceda1768b8dba08ab97