Revert "Don't run sync in upgrades"
Swift fails to deploy when using >1 storage node. The fact `need_sync`
was being set using information from ansible_local in an effort to
support rolling uprades sometime in the future. Sadly this variable
causes storage nodes to be removed from the key sync which results in a
failed build. If we decide serialized rolling uprades are required for
swift we can revisit this however until then we need to remove this
constraint and ensure we're allowing successfule deployments and
upgrades of swift.

This reverts commit 176452bebf7e1ceabd2cf72a89c558c0fc7f1420.

Change-Id: I21318d7c737c2415ef760b46176c3f9b372bb634