fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Change-Id: I6a2c6bbaa09b220c2b44e2a8d9be6a9a0bfa3453
Signed-off-by: Doug Hellmann <doug@doughellmann.com>